+++ src/test/java/lotto/domain/RandomNumberTicketFactoryTest.java
@@ -1,6 +1,7 @@
-package lotto.domain.ticketfactory;
+package lotto.domain;
 
 import lotto.domain.LottoTicket;
+import lotto.domain.RandomNumberTicketFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
